requirements.txt
attrs==16.0.0
bravado==8.3.0
bravado-core==4.3.0
cffi==1.7.0
crochet==1.5.0
cryptography==1.4
enum34==1.1.6
fido==3.2.0
functools32==3.2.3.post2
idna==2.1
ipaddress==1.0.16
jsonschema==2.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.0.0
python-dateutil==2.5.3
PyYAML==3.11
requests==2.11.0
rfc3987==1.3.6
service-identity==16.0.0
simplejson==3.8.2
six==1.10.0
strict-rfc3339==0.7
swagger-spec-validator==2.0.2
Twisted==16.3.0
webcolors==1.5
yelp-bytes==0.3.0
yelp-encodings==0.1.3
zope.interface==4.2.0
-e .
